DogeCoinV2
DogecoinV2 (DOGEV2) is a cryptocurrency and probably one of the most curious examples of digital tokens that exist. The cryptocurrency has a singular history: it was created after a meme in 2013 and everyone, including its founder, expected the joke token to disappear within a few weeks. Eight years later, the token is still one of the most popular and is often among the 30 cryptocurrencies by market capitalization.